This reverts commit 270d31c41c310b3422997f4849db4957bf24c2df.

+++ src/GatingHierarchy.cpp
@@ -344,7 +344,7 @@ void GatingHierarchy::transforming(bool updateCDF)
 			if(curTrans->gateOnly())
 				continue;
 
-			valarray<float> x(this->fdata.subset(curChannel));
+			valarray<double> x(this->fdata.subset(curChannel));
 			if(dMode>=GATING_HIERARCHY_LEVEL)
 				cout<<"transforming "<<curChannel<<" with func:"<<curTrans->getChannel()<<endl;
