+++ src/workspace.cpp
@@ -32,7 +32,7 @@ workspace::~workspace(){
 }
 
 
-valarray<float> workspace::toArray(string sCalTable){
+valarray<double> workspace::toArray(string sCalTable){
 
 
 //	sCalTable="3980,1.9428805e5,3981,1.9478264e5,3982,1.9527849e5,3983,1.9577559e5,3984,1.9627398e5";
@@ -40,7 +40,7 @@ valarray<float> workspace::toArray(string sCalTable){
 	boost::split(stringVec,sCalTable,boost::is_any_of(","));
 //
 
-	valarray<float> res(stringVec.size());
+	valarray<double> res(stringVec.size());
 	for(unsigned i=0;i<stringVec.size();i++)
 	{
 		res[i]=atof(stringVec.at(i).c_str());
